requirements.txt
python-dotenv==0.19.0
requests==2.26.0
python-telegram-bot==13.10

